chore: Move LoadingContainer to atoms
