+++ src/components/atoms/loading-container/index.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import Spinner from "../atoms/spinner"
+import Spinner from "../spinner"
 
 type LoadingContainerProps = {
   isLoading: boolean
