+++ src/domain/pricing/pricing-form/sections/product-prices.tsx
@@ -6,7 +6,7 @@ import EditIcon from "../../../../components/fundamentals/icons/edit-icon"
 import PlusIcon from "../../../../components/fundamentals/icons/plus-icon"
 import SearchIcon from "../../../../components/fundamentals/icons/search-icon"
 import TrashIcon from "../../../../components/fundamentals/icons/trash-icon"
-import LoadingContainer from "../../../../components/loading-container"
+import LoadingContainer from "../../../../components/atoms/loading-container"
 import { ActionType } from "../../../../components/molecules/actionables"
 import InputField from "../../../../components/molecules/input"
 import Modal from "../../../../components/molecules/modal"
